requirements.txt
requests==2.23.0
speedtest-cli==2.1.2